test(outboundapp): fold the model tests into the sibling test file
The tests only touch OutboundAppModel and Values, both exported, so they did
not need the package internal test package that forced them into a file of
their own.
